challenges/faust-data-processing/requirements.txt
faust-streaming==0.8.1
faker==13.2.0
prettyconf==2.2.1
